// src/interfaces/task.interface.ts
export interface Task {
  id: number;
  title: string;
  autor: string;
  statusTask: string;
  description: string;
  startDate: string;
  endDate: string;
}

export interface FormData {
  title: string;
  autor: string;
  statusTask: string;
  description: string;
  startDate: string;
  endDate: string;
}